Struct num_bigfloat::BigFloat

pub struct BigFloat { /* private fields */ }

Number representation.

Implementations

impl BigFloat

pub fn new() -> Self

Returns a new BigFloat with the value of zero.

pub fn from_bytes(bytes: &[u8], sign: i8, exponent: i8) -> Self

Creates a BigFloat value from a sequence of bytes. Each byte must represent a decimal digit. First byte is the most significant. bytes can be of any length. If bytes is longer than required, then the remaining bytes are ignored. If the “sign” is negative, then the resulting BigFloat will be negative.

Examples

let n1 = BigFloat::from_bytes(&[1,2,3,4,2,0,0,0], 1, -3);
let n2 = BigFloat::from_u32(12342);
assert!(n1.cmp(&n2) == Some(0));

pub fn from_f64(f: f64) -> Self

Creates a BigFloat from f64. The conversion is not guaranteed to be lossless since BigFloat and f64 have different bases.

pub fn from_f32(f: f32) -> Self

Creates a BigFloat from f32. The conversion is not guaranteed to be lossless since BigFloat and f64 have different bases.

pub fn to_f64(&self) -> f64

Converts BigFloat to f64.

pub fn to_f32(&self) -> f32

Converts BigFloat to f32.

pub fn to_i64(&self) -> Option<i64>

Converts BigFloat to i64. The function retursn None if self is Inf, NaN, or out of range of i64.

pub fn to_i128(&self) -> Option<i128>

Converts BigFloat to i128. The function retursn None if self is Inf, NaN, or out of range of i128.

pub fn to_u64(&self) -> Option<u64>

Converts absolute value of self to u64. The function retursn None if self is Inf, NaN, or out of range of u64.

pub fn to_u128(&self) -> Option<u128>

Converts absolute value of self to u128. The function retursn None if self is Inf, NaN, or out of range of u128.

pub fn get_mantissa_bytes(&self, bytes: &mut [u8])

Returns the mantissa of the BigFloat in bytes. Each byte represents a decimal digit. The first byte is the most significant. bytes can be of any length. If the length of bytes is less than the number of decimal positions filled in the mantissa, then the rest of the mantissa will be omitted.

The length of the mantissa can be determined using get_mantissa_len. If self is Inf or NaN, nothing is returned.

Examples

let n = BigFloat::from_f64(123.42);
let mut m = [0; 40];
n.get_mantissa_bytes(&mut m);
// compare m[0..10] to [1,2,3,4,2,0,0,0,0,0]
assert!(m[0..10].iter().zip([1,2,3,4,2,0,0,0,0,0].iter()).filter(|x| { x.0 != x.1 }).count() == 0);

pub fn get_mantissa_len(&self) -> usize

Returns the number of decimal places filled in the mantissa. If self is Inf or NaN, 0 is returned.

pub fn get_sign(&self) -> i8

Returns 1 if BigFloat is positive, -1 otherwise. If self is NaN, 0 is returned.

pub fn get_exponent(&self) -> i8

Returns the exponent part of the number. If self is Inf or NaN, 0 is returned.

pub fn set_exponent(&mut self, e: i8)

Sets the exponent part of the number. The function has no effect on Inf and NaN values.

pub fn to_raw_parts(&self) -> Option<([i16; 10], i16, i8, i8)>

Returns the raw parts of the number: the mantissa, the number of decimal places in the mantissa, the sign, and the exponent. If self is Inf or NaN, None is returned.

pub fn from_raw_parts( mantissa: [i16; 10], mantissa_len: i16, sign: i8, exponent: i8 ) -> Self

Creates a BigFloat from the raw parts. to_raw_parts can be used to get the raw parts of a number.

pub fn is_inf_pos(&self) -> bool

Returns true if self is positive infinity.

pub fn is_inf_neg(&self) -> bool

Returns true if self is negative infinity.

pub fn is_inf(&self) -> bool

Returns true if self is infinite.

pub fn is_nan(&self) -> bool

Return true if self is not a number.

pub fn add(&self, d2: &Self) -> Self

Adds d2 to self and returns the result of the addition.

pub fn sub(&self, d2: &Self) -> Self

Subtracts d2 from self and return the result of the subtraction.

pub fn mul(&self, d2: &Self) -> Self

Multiplies self by d2 and returns the result of the multiplication.

pub fn div(&self, d2: &Self) -> Self

Divides self by d2 and returns the result of the division.

pub fn cmp(&self, d2: &BigFloat) -> Option<i16>

Compares self to d2. Returns positive if self > d2, negative if self < d2, zero if self == d2, None if self or d2 is NaN.

pub fn inv_sign(&self) -> BigFloat

Reverses the sign of a number.

pub fn pow(&self, d1: &Self) -> Self

Returns self to the power of d1.

pub fn log(&self, b: &Self) -> Self

Returns the logarithm base b of a number.

pub fn is_positive(&self) -> bool

Returns true if self is positive. The function returns false if self is NaN.

pub fn is_negative(&self) -> bool

Returns true if self is negative. The function returns false if self is NaN.

pub fn is_subnormal(&self) -> bool

Returns true if self is subnormal. A number is considered subnormal if not all digits of the mantissa are used, and the exponent has the minimum possible value.

pub fn is_zero(&self) -> bool

Returns true if self is zero.

pub fn clamp(&self, min: &Self, max: &Self) -> Self

Restricts the value of self to an interval determined by the values of min and max. The function returns max if self is greater than max, min if self is less than min, and self otherwise. If either argument is NaN or min is greater than max, the function returns NaN.

pub fn max(&self, d1: &Self) -> Self

Returns the value of d1 if d1 is greater than self, or the value of self otherwise. If either argument is NaN, the function returns NaN.

pub fn min(&self, d1: &Self) -> Self

Returns value of d1 if d1 is less than self, or the value of self otherwise. If either argument is NaN, the function returns NaN.

pub fn signum(&self) -> Self

Returns a BigFloat with the value -1 if self is negative, 1 if self is positive, zero otherwise. The function returns NaN If self is NaN.

pub fn parse(s: &str) -> Option<Self>

Parses a number from the string s. The function expects s to be a number in scientific format in base 10, or +-Inf, or NaN.

Examples

use num_bigfloat::BigFloat;

let n = BigFloat::parse("0.0").unwrap();
assert!(n.to_f64() == 0.0);
let n = BigFloat::parse("1.123e-123").unwrap();
assert!((n.to_f64() - 1.123e-123).abs() < f64::EPSILON);
let n = BigFloat::parse("-Inf").unwrap();
assert!(n.to_f64() == f64::NEG_INFINITY);
let n = BigFloat::parse("NaN").unwrap();
assert!(n.to_f64().is_nan());

pub fn random_normal(exp_from: i8, exp_to: i8) -> Result<Self, Error>

Returns a random normalized (not subnormal) BigFloat number with exponent in the range from exp_from to exp_to inclusive. The sign can be positive and negative. Zero is excluded. Function does not follow any specific distribution law. The intended use of this function is for testing.

Errors

InvalidArgument - when exp_from is greater than exp_to.

pub fn classify(&self) -> FpCategory

Returns category of self as defined by core::num::FpCategory.

pub fn rem(&self, d1: &Self) -> Self

Returns the remainder of division of self by d1.

impl BigFloat

pub fn abs(&self) -> Self

Returns the absolute value of self.

pub fn int(&self) -> Self

Returns the integer part of self.

pub fn frac(&self) -> Self

Returns the fractional part of self.

pub fn ceil(&self) -> Self

Returns the smallest integer greater than or equal to self.

pub fn floor(&self) -> Self

Returns the largest integer less than or equal to self.

pub fn round(&self, n: usize, rm: RoundingMode) -> Self

Returns a rounded number with n decimal positions in the fractional part of the number using the rounding mode rm.

pub fn sqrt(&self) -> Self

Returns the square root of self.

pub fn cbrt(&self) -> Self

Returns the cube root of self.

pub fn ln(&self) -> Self

Returns the natural logarithm of self.

pub fn log2(&self) -> Self

Returns the logarithm base 2 of self.

pub fn log10(&self) -> Self

Returns the logarithm base 10 of self.

pub fn exp(&self) -> Self

Returns e to the power of self.

pub fn sin(&self) -> Self

Returns the sine of self. The function takes an angle in radians as an argument.

pub fn cos(&self) -> Self

Returns the cosine of self. The function takes an angle in radians as an argument.

pub fn tan(&self) -> Self

Returns the tangent of self. The function takes an angle in radians as an argument.

pub fn asin(&self) -> Self

Returns the arcsine of self. The result is an angle in radians ranging from -pi/2 to pi/2.

pub fn acos(&self) -> Self

Returns the arccosine of self. The result is an angle in radians ranging from 0 to pi.

pub fn atan(&self) -> Self

Returns the arctangent of self. The result is an angle in radians ranging from -pi/2 to pi/2.

pub fn sinh(&self) -> Self

Returns the hyperbolic sine of self.

pub fn cosh(&self) -> Self

Returns the hyperbolic cosine of self.

pub fn tanh(&self) -> Self

Returns the hyperbolic tangent of self.

pub fn asinh(&self) -> Self

Returns the inverse hyperbolic sine of self.

pub fn acosh(&self) -> Self

Returns the inverse hyperbolic cosine of self.

pub fn atanh(&self) -> Self

Returns the inverse hyperbolic tangent of self.

impl BigFloat

pub fn from_i8(i: i8) -> Self

Construct BigFloat from integer value.

pub fn from_u8(i: u8) -> Self

Construct BigFloat from integer value.

impl BigFloat

pub fn from_i16(i: i16) -> Self

Construct BigFloat from integer value.

pub fn from_u16(i: u16) -> Self

Construct BigFloat from integer value.

impl BigFloat

pub fn from_i32(i: i32) -> Self

Construct BigFloat from integer value.

pub fn from_u32(i: u32) -> Self

Construct BigFloat from integer value.

impl BigFloat

pub fn from_i64(i: i64) -> Self

Construct BigFloat from integer value.

pub fn from_u64(i: u64) -> Self

Construct BigFloat from integer value.

impl BigFloat

pub fn from_i128(i: i128) -> Self

Construct BigFloat from integer value.

pub fn from_u128(i: u128) -> Self

Construct BigFloat from integer value.